the MkIntegerTerm macro really, really requires Int arguments
git-svn-id: https://yap.svn.sf.net/svnroot/yap/trunk@214 b08c6af1-5177-4d33-ba66-4b1c6b8b522a

+++ C/stdpreds.c
@@ -1369,7 +1369,7 @@ cont_current_predicate(void)
   else
     name = (Atom)pp->FunctorOfPred;
   return (unify(ARG2,MkAtomTerm(name)) &&
-	  unify(ARG3, MkIntegerTerm(Arity)));
+	  unify(ARG3, MkIntegerTerm((Int)Arity)));
 }
 
 static Int
